neutrino: fix 32<->64bit format string warnings
use portable C99 format string macros for 64bit types to
fix many warnings when compiling for 64bit architectures,
add some (int) casts for size_t

+++ src/zapit/src/bouquets.cpp
@@ -389,7 +389,7 @@ void CBouquetManager::parseBouquetsXml(const char *fname, bool bUser)
 				newBouquet->sortBouquet();
 			search = search->xmlNextNode;
 		}
-		INFO("total: %d bouquets", Bouquets.size());
+		INFO("total: %d bouquets", (int)Bouquets.size());
 	}
 	xmlFreeDoc(parser);
 }
